Move utils to utils_test.go file
